+++ bromine/reversi/__init__.py
@@ -1,0 +1,3 @@
+from .sys import reversi_sys
+
+__all__ = ["reversi_sys"]
